modified deploy/osx.sh to make possible to run app directly from sources

+++ src/deploy/osx.sh
@@ -5,16 +5,24 @@
 CURRENT_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
 
 ROOT_DIR="$HOME/.bitdust"
+
 LOG_FILE="${ROOT_DIR}/install.log"
+
 SOURCE_DIR="${ROOT_DIR}/src"
 VENV_DIR="${ROOT_DIR}/venv"
+
 BITDUST_PY="${SOURCE_DIR}/bitdust.py"
 BITDUST_COMMAND_FILE="${ROOT_DIR}/bitdust"
-GLOBAL_COMMAND_LOCATION="/usr/local/bin"
-GLOBAL_COMMAND_FILE="${GLOBAL_COMMAND_LOCATION}/bitdust"
+
 PYTHON_BIN="${ROOT_DIR}/python/bin/BitDust-p2p-app"
 PYTHON_VENV_BIN="${ROOT_DIR}/venv/bin/BitDust-node"
 PIP_VENV_BIN="${ROOT_DIR}/venv/bin/pip"
+
+
+if [[ ! -f $PYTHON_BIN ]]; then
+    PYTHON_BIN="python3"
+    PYTHON_VENV_BIN="${ROOT_DIR}/venv/bin/python3"
+fi
 
 
 if [[ "$1" == "stop" ]]; then
@@ -61,19 +69,23 @@ fi
 cd "$ROOT_DIR"
 
 if [[ ! -e $SOURCE_DIR ]]; then
-    echo ''
-    echo "##### downloading source code files from Git repository"
-    mkdir -p "$SOURCE_DIR"
 
-    if [[ ! -f $PYTHON_VENV_BIN ]]; then
+    if [[ ! -f $BITDUST_COMMAND_FILE ]]; then
         echo ''
-        echo "##### creating Python virtual environment"
-        $PYTHON_BIN -m venv $VENV_DIR
-        cp "${ROOT_DIR}/venv/bin/python3" "${PYTHON_VENV_BIN}"
-        $PIP_VENV_BIN install -q --upgrade pip || (echo "pip upgrade failed" && exit 1)
+        echo "##### downloading source code files from Git repository"
+        mkdir -p "$SOURCE_DIR"
+
+        if [[ ! -f $PYTHON_VENV_BIN ]]; then
+            echo ''
+            echo "##### creating Python virtual environment"
+            $PYTHON_BIN -m venv --clear --copies $VENV_DIR
+            cp "${ROOT_DIR}/venv/bin/python3" "${PYTHON_VENV_BIN}"
+            $PIP_VENV_BIN install -q --upgrade pip || (echo "pip upgrade failed" && exit 1)
+        fi
+
+        $PYTHON_BIN -c "import pygit2; pygit2.clone_repository('https://github.com/bitdust-io/public.git', '$SOURCE_DIR')" || (echo "git clone failed" && exit 1)
     fi
 
-    $PYTHON_BIN -c "import pygit2; pygit2.clone_repository('https://github.com/bitdust-io/public.git', '$SOURCE_DIR')" || (echo "git clone failed" && exit 1)
 else
     echo ''
     echo "##### updating source files from Git repository"
@@ -81,7 +93,7 @@ else
     if [[ ! -f $PYTHON_VENV_BIN ]]; then
         echo ''
         echo "##### creating Python virtual environment"
-        $PYTHON_BIN -m venv $VENV_DIR
+        $PYTHON_BIN -m venv --clear --copies $VENV_DIR
         cp "${ROOT_DIR}/venv/bin/python3" "${PYTHON_VENV_BIN}"
         $PIP_VENV_BIN install -q --upgrade pip || (echo "pip upgrade failed" && exit 1)
     fi
@@ -92,26 +104,25 @@ fi
 
 if [[ ! -e $PIP_VENV_BIN ]]; then
     echo ''
-    echo "##### installing Python packages"
-    $PIP_VENV_BIN --default-timeout=10 install -U -q -r "$SOURCE_DIR/requirements.txt" || (echo "pip requirements install failed" && exit 1)
-    echo "#!/bin/sh" > $BITDUST_COMMAND_FILE
-    echo "$PYTHON_VENV_BIN $ROOT_DIR/src/bitdust.py \"\$@\"" >> $BITDUST_COMMAND_FILE
-    chmod +x $BITDUST_COMMAND_FILE
 else
     # TODO: this is slow and can fail if user is offline...
     # this actually must be only executed when requirements.txt was changed
-    echo ''
-    echo "##### updating Python packages"
-    $PIP_VENV_BIN --default-timeout=10 install -U -q -r "$SOURCE_DIR/requirements.txt" || (echo "pip requirements install failed" && exit 1)
-    echo "#!/bin/sh" > $BITDUST_COMMAND_FILE
-    echo "$PYTHON_VENV_BIN $ROOT_DIR/src/bitdust.py \"\$@\"" >> $BITDUST_COMMAND_FILE
-    chmod +x $BITDUST_COMMAND_FILE
+    if [[ -f $SOURCE_DIR/requirements.txt ]]; then
+        echo ''
+        echo "##### updating Python packages"
+        $PIP_VENV_BIN --default-timeout=10 install -U -q -r "$SOURCE_DIR/requirements.txt" || (echo "pip requirements install failed" && exit 1)
+    fi
 fi
 
 
 echo ''
 echo '##### starting main process in background'
-$PYTHON_VENV_BIN $BITDUST_PY daemon
+
+if [[ ! -f $BITDUST_COMMAND_FILE ]]; then
+    $PYTHON_VENV_BIN $BITDUST_PY daemon
+else
+    $BITDUST_COMMAND_FILE daemon
+fi
 
 
 echo ''
